Backend/requirements.txt
Flask
flask-cors
requests
pymongo
python-dotenv
gunicorn
langgraph
